+++ services/cost_management/cost_management_service/src/main/java/org/changppo/cost_management_service/service/member/oauth2/OAuth2Client.java
@@ -1,8 +1,8 @@
-package org.changppo.cost_management_service.service.member.oauth;
+package org.changppo.cost_management_service.service.member.oauth2;
 
 import java.io.IOException;
 
-public interface OAuth2Service {
+public interface OAuth2Client {
     void unlinkMember(String providerMemberId) throws IOException;
     boolean supports(String provider);
 }
